Move Forms::DeleteConfirmationInput to DeleteConfirmationInput
