+++ app/input_objects/delete_confirmation_input.rb
@@ -1,0 +1,5 @@
+class DeleteConfirmationInput < ConfirmActionInput
+  def to_partial_path
+    "input_objects/delete_confirmation_input"
+  end
+end
