+++ app/input_objects/forms/delete_confirmation_input.rb
@@ -1,5 +1,2 @@
-class Forms::DeleteConfirmationInput < ConfirmActionInput
-  def to_partial_path
-    "input_objects/forms/delete_confirmation_input"
-  end
+class Forms::DeleteConfirmationInput < DeleteConfirmationInput
 end
