+++ app/views/input_objects/_delete_confirmation_input.html.erb
