+++ config/locales/input_objects/delete_confirmation.yml
@@ -3,7 +3,7 @@ en:
   activemodel:
     errors:
       models:
-        forms/delete_confirmation_input:
+        delete_confirmation_input:
           attributes:
             confirm:
               blank: You must choose an option
