+++ spec/views/pages/delete.html.erb_spec.rb
@@ -33,7 +33,7 @@ RSpec.describe "pages/delete" do
   end
 
   it "has a delete confirmation input to confirm deletion of the page" do
-    expect(rendered).to render_template "input_objects/forms/_delete_confirmation_input"
+    expect(rendered).to render_template "input_objects/_delete_confirmation_input"
   end
 
   describe "delete confirmation input" do
